requirements.txt
tensorflow
numpy
opencv-python
pandas
pillow
customtkinter